Update patch set 3
Patch Set 3:

Apologies for the peanut gallery comment, but would you mind submitting a CL that applies this very change to the https://godoc.org/golang.org/x/tools/go/vcs package as well?

That package is an exported copy of this functionality, and many Go tools that want to perform import path resolution the same way as `go get` does use it.

Because issue https://github.com/golang/go/issues/11490 is not yet resolved we have to manually apply changes to both packages to keep them in sync.

It's completely okay if you don't want to do this, in that case I will try to do it myself, just let me know... Thanks.

Patch-set: 3