Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> In the past we've made an effort to preserve both names and semantics of benchmarks to allow easy, meaningful cross-version comparisons. I like the code better now but I'm not sure it's worth losing comparability.

Marcel brought this up in one of the CLs (not this one). I think the cleanup here is worth the breakage. Really benchmarks being run across versions from different source code are a little suspect anyway. I know we do it, but if it's a little harder to do automatically, that's OK with me. Like in other aspects, we'd like the standard library to demonstrate best practices, and these are much cleaner than before and good real-world examples to point people at.

Fixing the names for a cross-version comparison is a simple sed script, and it's a one-time split.

Patch-set: 2