Update patch set 12
Patch Set 12:

> Patch Set 12:
> 
> > (2 comments)
>  > 
>  > Any idea why some of the small memmoves are slower?  It doesn't
>  > make much sense to me.  Memmove8 is +~10%, Memmmove9 is -~10%, and
>  > nothing in this CL changed any instruction that either of those
>  > benchmarks run.
> 
> My understanding is that they are results of optimization. If you look on code generated you see that Go's assembler reorders basic blocks now. I've reported the problem to Russ Cox - I hope he's investigating the problem.

So I talked to Russ, Austin, and gang and we think it would be better for this CL to enable this optimized copy any time it is correct: enabledAVX && hasFeature(cfAVX), right?  Then we can exempt particular processors which we know have those instructions but don't perform well with them.

> On the SandyBridge your old implementation with backward REP MOVSQ can be 2 times faster with unaligned 2K buffers. Fast unaligned loads works well on Haswell and newer processors.

So is there a way to detect SandyBridge?  Maybe just check for AVX2?  Or maybe we just don't care about backward unaligned copies.  Go rarely triggers this case.  It happens only with explicit user copy() calls (as opposed to forward copies which happen all the time, e.g. for append() calls).

Patch-set: 12