Update patch set 4
Patch Set 4:

(1 comment)

> I can confirm that this change fixes the test on yakkety (testsanitizers now fails though -- need to report another bug about that :/)
> 
> I am a little confused though, it seems to me that there are two changes here: cmd/pprof's change to use the expected vs actual load address by looking at segment headers, and the change to record /proc/self/maps. Are both required? I don't really see how they interact.

Thanks for testing it.

Both changes are required.  Writing out the mapping data is needed so that pprof can map the PC to a shared library (or, in this case, a PIE, which is just a shared library really).  pprof already knows how to do that, as the legacy profiling formats supports memory maps in the GNU/Linux /proc/self/maps format.

The second change is required so that the Go specific code in cmd/pprof/pprof.go will take the load address, already being passed in, and use it when mapping PC values to function/file/line information in the Go pclntab.

Patch-set: 4