Update patch set 3
Patch Set 3:

I am confused by this.  The test says

// Trailing .git is less preferred but included for
// compatibility purposes while the same source needs to
// be compilable on both old and new go

But the actual code sets check to noVCSSuffix, which does not accept the trailing ".git".  This was not noticed because the test is not run when using -test.short, so basically the test has always failed.

What is the intent here?  Since nobody seems to want the trailing .git, and since the trailing .git does not in fact work, should we just change the test?

This caused issue #15979.

Patch-set: 3