Update patch set 2
Patch Set 2:

> R=go1.8
 > 
 > This needs a test.
 > 
 > It's also not obvious to me that we want to support this.  Perhaps
 > we do, but in general we have no symlink support in GOPATH.  Adding
 > any requires tests and docs.

The purpose of this patch isn't to add any symlink support in GOPATH, but rather to keep the tools from puking if the current working directory contains a symlink.  My GOPATH does not have symlinks, rather I'm trying to run 'go fix/fmt/vet' from a working directory that does contain symlinks, e.g. 'go fmt ./...'.

I understand not wanting to support symlinks as it opens a can of worms, but relPaths' implementation and the tools' usage of relPaths feels like they are going out of their way to not work with even the most basic usage of symlinks.  If relPaths didn't exist then the tools would not break in this scenario, albeit the output would be more verbose in many scenarios due to not using relative paths.

I'll get working on a test(s).

Patch-set: 2