Update patch set 2
Patch Set 2:

> > > Doesn't TestParseCanned fail?
 > >
 > > that's why parser_test.go change is here as well - supplying some
 > > bin name.
 > 
 > But it passes a wrong binary. Current binary is not the binary that
 > produced the canned traces. It should have been failed...

I see. 

What happened is the symbolize function doesn't fail as long as any readable binary file is given. If the pc is invalid, addr2line spits out ? or ?:0, so the worst case, the pc info is filled with bunch of garbages. There is no sanity check on the binary version.

Thankfully the new format is self-contained. 

Updated the test. At least, I want the trace command to report an error message in case users misuse the command.

Patch-set: 2