cmd/internal/obj: reduce per-architecture opcode space
DO NOT REVIEW

[code freeze]

s390x took up the last available chunk of int16 opcodes.
There is a RISC-V port in progress out of tree,
and there will likely be other architectures.
Reduce the opcode space to allow more architectures to
fit without increasing to int32.

This is the smallest power of two that accomodates all
existing architectures. All else being equal, smaller is
better--smaller numbers are easier to generate immediates
for and easier on the eyes when debugging.

Change-Id: I4d0824b28913892fbd0579d3f90bea34e44c8946