Update patch set 1
Patch Set 1:

R=go1.8

I'm concerned about putting it in the reflect package because as far as I can see that would mean adding a lock, and I can't predict how that would affect overall behavior.

I don't know if this is always a good idea or not, either, so it probably should not go into 1.7.

Patch-set: 1