runtime: over-assist in terms of scan work instead of allocation bytes
DO NOT REVIEW. Experimenting for #14812.

Currently we over-assist to build up 1MB of extra allocation credit.
However, how much extra scan work this translates into depends on the
assist ratio; if the ratio is high, this could mean a significant
amount of extra time spent scanning to build up this credit, which
introduces latency.

Instead, over-assist to build up 512K of scan work credit. This is
much more directly related to how much extra time we spend scanning to
build up this credit, so it spreads out assists better when the assist
ratio is high.

A possibly better alternative would be to eliminate the over-assist
mechanism entirely and instead couple assists with acquiring a span
into the mcache (charging for all free slots in the acquired span).
This would naturally amortize the cost of assists, without the danger
of under-assisting.

Change-Id: Ic5501ad8c2d64af30a54ead23ff6b85def9b126b