runtime: adjust trigger based only on CPU utilization
DO NOT REVIEW. Experimenting for #14812.

Currently we adjust the trigger based on a combination of CPU
utilization and heap overshoot. However, this has some problems.
Assists also control for heap overshoot, which means the trigger
controller is fighting with assists for control over heap overshoot.
Assists also make it essentially impossible to overshoot, which means
that the trigger controller only gets a one-sided signal on heap size.
It also only gets a one-sided signal on CPU utilization: since we fix
the mark worker utilization at 25%, CPU utilization can only ever be
too high. Finally, since the trigger controller depends two variables
but only controls one, it has a isocline that results in no adjustment
if CPU utilization is too high and the heap undershoots. Because of
the earlier problems, this is where we almost always settle.

This is an attempt to fix this by changing the trigger error formula
to account only for CPU utilization. This leaves control of heap
overshoot to the assists, and means the trigger controller has a
single input and a single output and hence doesn't have the
problematic isocline of the current controller. We also force the
controller to target a slightly higher CPU utilization than we pin the
mark workers at so that it has a two-sided signal on CPU utilization.

Change-Id: Ib6a7e16882088e5bf8138cd367987f28eca4654a