runtime: shorten mark worker preemption to 1ms
DO NOT REVIEW. Experimenting for #14812.

Currently fractional and idle mark workers run until they run out of
work or until they are preempted by the usual 10ms preemption
mechanism. But 10ms is a rather long time to hog to CPU for some
applications.

This adds an explicit self-preemption after 1ms for fractional and
idle mark workers.

Even better would be for idle workers to detect when there's other
work to do and immediately return to the scheduler, but that requires
mechanisms in the scheduler that don't exist today.

Based on CL 21036.

Change-Id: I883ed3960537482b199bda80079cdc6e287fe616