runtime: essentially disable dedicated and fractional mark workers
DO NOT REVIEW. Experimenting for #14812.

This is an experiment with letting the assists do essentially all GC
work.

This does lead to dramatically better interleaving in allocation-heavy
workloads during GC. However, it doesn't seem to help latency much.
This may be because of the following two problems:

1) Assists can't currently do root marking jobs, and hence all of the
assists that run before the idle worker has kicked in will block
immediately.

2) Blocked assists currently can only be woken up by mark workers,
which means they could be blocked until the end of the GC cycle, even
though assists that started much later are finishing quickly. This
behavior is fundamentally unfair, and suggests that we may want to
process all assists in FIFO order and disallow this sort of "barging"
behavior they currently exhibit. We could do this by having a global
notion of the scan work and the heap allocation and, on assist,
computing the global scan work goal for the current heap allocation
and simply assisting until we've reach that scan work goal. Assists
would hence complete in FIFO order, but we wouldn't have the current
artificial coupling of the allocating goroutine and the scanning
goroutine that's constraining the system.

Change-Id: Ied7396b8e8feea31b215b38bb078d54b08ef1a0e