Update patch set 1
Patch Set 1:

> The only thing to double-check is that zeroing is guaranteed to happen before heap allocation of escaping paramouts.

As far as I can tell, it does.  The zeroing happens after code generation for the function.  The code generation for the function includes the heap allocation (heapmoves called by walk).  The zeroing is inserted as the very first instructions of the function (by Defframe), so it is executed before the heap allocation.

Also verified by looking at the generated code.

Patch-set: 1