cmd/compile: optimize string comparison with string literals
Especially for short string literals, calling an external function
is unnecessary overhead. The special case of s == "" is already
handled. This change improves comparison with string literals of
up to 8 bytes by comparing them bytewise.

When rewriting the tests in walk_test.go to be benchmarks, there is an
improvement from 275 ns/op to 236 ns/op. The overall speed improvement
in typical programs will not be noticeable, but at the same time, the
size of the generated binaries is reduced. The size of the go.exe file
went down from 9858048 bytes to 9839104, being reduced by 18944 bytes.
Furthermore, when the optimizer someday propagates string literals when
inlining function calls, this will benefit calls to strings.HasPrefix
and strings.HasSuffix as well.

When running make.bat with this change, logging which strings get
optimized, it reports 2284 instances where this optimization is applied,
most of them being the strings “init” (185 times) and “_” (178 times),
followed by platform names like “arm”, “amd64” (each 50 times). This
optimization is more effective than the "badgerbadgerbadger" one; the
latter is called 0 times during make.bat).

Change-Id: I6e7fa1e2adf99774fcab4a8b5f52a782615841f5