Update patch set 3
Patch Set 3:

Regarding the benchmarks: I just modified my local walk_test.go to be a benchmark, and it shows an improvement from 275 ns/op to 236 ns/op.

This change probably has an effect on the standard library, but definitely on the Go compiler itself. The size of the go.exe file  went down from 9858048 bytes to 9839104, being reduced by 18944 bytes. The speedup won’t be noticeable in large programs, but the change in code size benefits everyone.

Furthermore, when the optimizer someday propagates string literals when inlining function calls, this will benefit calls to strings.HasPrefix and strings.HasSuffix as well. This is the optimization I originally hoped for, since I use these functions a lot when parsing text files.

Patch-set: 3