Update patch set 1
Patch Set 1:

It would be nice to figure out whole call trees which don't need checks because their total stack space is small enough (stackguard-stacksmall in total).  This would just be a subset of that.

The "leaves" of the trees don't actually need to be leaves, just that any call they make goes to a function which does a stack check.  I think this CL is a subset of that rule.

Any function which could be the target of an indirect call would also need a stack check.

I don't know how we should compute these no-check trees.  Some sort of annealing starting at one-function trees might work.

Patch-set: 1