Update patch set 1
Patch Set 1:

This is one approach to fixing #16378. The alternative is to teach cmd/go not to set the install suffix for default for these platforms. I'm slightly inclined to this approach, because the -shared flag is in fact set during compilation, but I don't really have a strong preference.

Patch-set: 1