Update patch set 1
Patch Set 1:

The problem I see with changing the install suffix by default is that it crops again in other places, such as misc/cgo/testcarchive/carchive_test.go.  The install suffix is useful for separating code compiled in one way from code compiled in a different, incompatible, way.  For darwin/arm and darwin/arm64 it seems that there is no way to compile code without -shared.  So the install suffix doesn't seem to be serving any purpose.

To put it another way, you seem to be suggesting that the install suffix is useflu on darwin/arm and darwin/arm64 because it indicates that the code was compiled with -shared.  But that fact is only meaningful if you want to distinguish from some other case, and it seems to me that there is no other case.

Perhaps we should simply move the -shared default for those targets from cmd/go to cmd/compile.

Patch-set: 1