Update patch set 1
Patch Set 1: -Run-TryBot

I see. I just started implementing that, but hit a few bumps, so I wanted to check in before continuing:

It's not just cmd/compile that needs the "default to -shared" list; gccgo and cmd/asm need it too.

The setting of -shared for some architectures (but not the same list) also occurs in non-default build modes, like c-archive, c-shared, and pie. I'm not sure whether those decisions should also be moved to the compilers/assemblers.

Patch-set: 1
Label: Run-TryBot=0