runtime: don't hard-code physical page size
DO NOT SUBMIT. For 1.8. Benchmarks.

Currently the runtime hard-codes a physical page size used by the OS
and, on Linux, double checks that the true physical page size is
compatible with the assumed physical page size.

Since we're getting the true physical page size from the OS anyway,
make the physical page size a variable instead of a constant.

Change-Id: I6c23b9df860db309c38c8287a703c53817754f03