runtime: don't hard-code physical page size
DO NOT SUBMIT. For 1.8. Benchmarks.

Now that the runtime fetches the true physical page size from the OS,
make the physical page size used by the allocated a variable instead
of a constant. This isn't used in any performance-critical paths, so
it shouldn't be an issue.

sys.PhysPageSize is also renamed to sys.DefaultPhysPageSize to make it
clear that it's not necessary the true page size. There are no uses of
this constant any more, but we'll keep it around for now.

Updates #12480 and #10180.

Change-Id: I6c23b9df860db309c38c8287a703c53817754f03