Update patch set 3
Patch Set 3:

thanks, this is nice.

 > I don't know any place in particular where it triggers.  But in
 > fmt, for example, we do:
 > 
 > next = u / 10
 > buf[i] = byte('0' + u - next*10)
 > u = next
 > 
 > We could now do
 > 
 > u, c = u/10, u%10
 > buf[i] = byte('0' + c)
 > 
 > with no loss in performance.  In fact, it might be faster.

Note that this is a case of division by a constant which the compiler usually transforms into a multiply shift which is far cheaper than a div. I could not get dev.ssa to emit a DIVQ for the above. When i replace the code to use a variable that is initialized to 10 (u, c = u/d, u%d) it performs worse: BenchmarkFprintInt increases from 155ns to 185ns. On master without the optimization the timing is around 218ns when using two DIVQ.

However we have the case of variable base transformation in https://github.com/golang/go/blob/master/src/strconv/itoa.go#122 where i would guess it will lead to an improvement after refactoring the code to trigger the optimization.

Patch-set: 3