[dev.garbage] runtime: publish a local object
ROC needs to publish objects before they are made
visible to other goroutines. Publishing an object
includes a transitive walk of all local objects
reachable from the object being published. While
this is a naturally recursive operation it needs
to be performed by code that has strict limits on
available stack size. For this reason the transitive
walk is done using work buffers similar to those
used by the garbage collector to mark the heap.

Change-Id: Iaa26c7dc6cd80e4e489c57ab36f065d15a304a44