[dev.garbage] runtime: ROC write barrier code
ROC (request oriented collector) employs a write barrier
that maintains an invariant that all published objects
have their allocation bit set. This CL implements the
code that maintains this invariant.

The high level approach is to examine the referenced
object and the referent object holding the slot. If the
referent has its allocation bit set and the reference
does not then installing the reference in the referent
is a publishing action. Before this can happen the
publishing code transitively traces the unpublished
object's references ensuring the are also published.
The straightforward recursive implementation grows
the stack at unfortunate times so we borrow much of
the mark logic buffer data structures to implement
a interative approach.

After all reachable objects have been marked as public
the reference is installed. It is important not to
install the reference until the unpublished objects have
had their mark bits set for two reasons:

1) Unpublished objects are local and marking them
   before they are published does not require
   synchronization with other goroutines. It is a
   totally local operation.

2) If we installed the reference before we did the
   transitive publishing then another goroutine could
   race ahead of the publisher, manipulate pointers,
   and break the algorithm.

This code is not needed or turned on at the moment
since it is only meaningful when used on concert with
the rest of the ROC code.

Change-Id: I747909432e65bb8417eb83333953581500eeb974