Update patch set 1
Patch Set 1:

Just chatted with Cherry and it sounds like this is currently a problem on MIPS, since the runtime is configured for 16k pages, but both 16k and 4k pages are common. Hence, while I think I will defer the follow-up to use a 64k page size on ARM, I still think this CL should go in for 1.7.

Patch-set: 1