cmd/compile: change phi location to be optimistic at backedges
This is:

one simple trick that cuts the number of phi-nodes
(temporarily) inserted into the ssa representation by a factor
of 10, and can cut the user time to compile tricky inputs like
gogo/protobuf tests from 13 user minutes to 9.5, and memory
allocation from 3.4GB to 2.4GB.

a this-time-for-sure sparse lookup, that does not rely on
an assumption proven false by at least one pathological
input "etldlen".

With the combination of these two changes, phi locations
for etldlen completes in 2 minutes instead of 20 or longer,
and without a gigantic memory footprint.

Updates #16407

Change-Id: Iaa8aaa8c706858b3d49de1c4865a7fd79e6f4ff7