Update patch set 3
Patch Set 3: Code-Review+1

(2 comments)

I'm afraid I'm confused by this code.  I'm not really sure I ever understood it.
I don't understand what the changes in sparsetreemap.go have to do with the changes in sparselocatephifunctions.go.  The former look safe enough for 1.7.  I don't understand the latter enough to say that.

The CL description says a lot about the effects of this change, but not enough about how the change does it.

Patch-set: 3
Label: Code-Review=+1