Update patch set 4
Patch Set 4:

I attempted to add enough commentary to the code to make it less confusing.
The two fixes are unrelated; etldlen tickles two separate bugs.

A sparse tree is a subset of a larger tree, containing only "def" nodes.
The representation of a sparse tree is (conceptually) a sorted list (actually, a balanced binary tree data structure) where each def node is represented by entry and exit sparse-tree nodes; effectively, parentheses.  The parentheses order is obtained by numbering left parentheses with the tree-walk entry number of a def block and right parentheses with the tree-walk exit number of a def block.  A query, for any node U within the larger tree (i.e., a "use") finds the greatest-lower-bound(not EQ) of U's entry number in the sparse tree.  If it is a left parentheses, then it directly brackets U (occurs directly above it in the sparse tree, is a def for U).  If it is a right parentheses R, then U and R are directly bracketed by the same sparse parent; find that sparse parent.

The modified structure maintains sparse parents directly, instead of repeatedly searching.  This requires an update at insertion into the tree.

Patch-set: 4